+++ tests/test_flows.py
@@ -335,3 +335,43 @@ class TestJobsRunsSubmitAndWaitForCompletion:
                     }
                 ],
             )
+
+    @pytest.mark.respx(assert_all_called=True)
+    async def test_run_success_missing_run_name(
+        self, common_mocks, respx_mock, databricks_credentials
+    ):
+        respx_mock.get(
+            "https://dbc-abcdefgh-123d.cloud.databricks.com/api/2.1/jobs/runs/get?run_id=36108",  # noqa
+            headers={"Authorization": "Bearer testing_token"},
+        ).mock(
+            return_value=Response(
+                200,
+                json={
+                    "state": {
+                        "life_cycle_state": "TERMINATED",
+                        "state_message": "",
+                        "result_state": "SUCCESS",
+                    },
+                    "tasks": [{"run_id": 36260, "task_key": "prefect-task"}],
+                },
+            )
+        )
+
+        respx_mock.get(
+            "https://dbc-abcdefgh-123d.cloud.databricks.com/api/2.1/jobs/runs/get-output",  # noqa
+            headers={"Authorization": "Bearer testing_token"},
+        ).mock(return_value=Response(200, json={"notebook_output": {"cell": "output"}}))
+
+        result = await jobs_runs_submit_and_wait_for_completion(
+            databricks_credentials=databricks_credentials,
+            tasks=[
+                {
+                    "notebook_task": {
+                        "notebook_path": "path",
+                        "base_parameters": {"param": "a"},
+                    },
+                    "task_key": "key",
+                }
+            ],
+        )
+        assert result == {"prefect-task": {"cell": "output"}}